Use a custom more robust - but slower - polygonizer
This is inspired by and based on our original `AreaTree` / `AreaMeta` construction from 1.14 and earlier. For each
clockwise ring, we construct a `GeometryUtil.Island` to represent the solid portion of the geometry. For each
counterclockwise ring, we treat it as an ocean. After retrieving all islands and oceans, we stitch each ocean to each
island as an interior polygonal ring. Once all oceans are connected to islands, we convert each island to a JTS
`Polygon`.

Differences from the original `AreaTree` construction are:
1. No need to build an entire tree. We only connect oceans to parent islands, but we do not care about connecting
   islands to parent oceans. Each island with their child oceans becomes a polygon.
2. The resulting polygons are run through a `GeometryPrecisionReducer` so callers will not need to worry about
   ultra-precise points. This reducer internally ensures the results remain valid, e.g., by eliminating mistaken
   self-intersections that would cause confusion about interior and exterior points.