Prep campaign properties UI for separated auras
The full set of lights is split into two separate collections on-demand: one for lights and one for auras. The lights
are displayed in the Light tab, and the auras are displayed in the Auras tab. On commit, both are merged back together
and committed to the campaign.

`AuraSyntax` is also separated out from `LightSyntax`, but is a carbon copy for now. Neither has been reduced to reject
inapplicable syntax yet.

For some cleanup, the sightPanel, lightPanel, auraPanel, sightHelp, lightHelp, and auraHelp components are bound to
fields in `CampaignPropertiesDialogView` so they can more conveniently be referenced without requiring
`AbeillePanel#getComponent()`.