Be more intentional about light source map structure
The big deal here is that no one is allowed to implicitly modify the campaign's light sources map. I.e., we no longer
return a mutable reference to the internal map. Instead, a copy is returned (unmodifiable map wouldn't be as nice since
we have two levels of maps to deal with). Changes must be explicitly written back using `Campaign#setLightSourcesMap()`
or `CampaignProperties#setLightSourcesMap()`.

The next change is a nice-to-have and requires the former change to properly achieve. The outer map of the light sources
map remains a `TreeMap<>`, but this is now notably enforced now via `readResolve()`. But the inner map has been changed
from a `HashMap<>` to a `LinkedHashMap<>` to preserve insertion order. The `CampaignPropertiesDialog`, `LightSyntax`,
and `AuraSyntax` also use `LinkedHashMap<>` now so that the user input is always preserved. This is a bit of a stop-gap
solution while we still depend on the current structure of the light sources map. In the future I expect to replace this
structure with something that does not depend on the same collection implementation being selected in a dozen different
places.