Specialize light and aura syntaxes
Light syntax is still represented by `LightSyntax`, while aura syntax is represented by `AuraSyntax`.

Compared to previously, aura syntax does not support lumens. Internally, the `Light` objects for auras will be set to
the default lumens value of 100. Note that with the current syntax, we cant *forbid* a lumens string from being
provided, because our regex is not properly anchored. This is also true of lights: any random characters can be appended
to the end of a range argument and they will simply be discarded without error or warning. Since this is not a new
problem, I've left it as is, but I hope it can be tightened up in the future so there is no confusion.

Lights no longer suppor the aura keyword, or the GM or OWNER arguments. Providing any of these will result in an error
since they are not recognized and will be interpreted as a range argument.