Make instruction and tooltip keys into constructor parameters for AbstractDrawingLikeTool
This fixes tooltips for drawing, expose and topology tools which are instance-based. This is because the `Tool` class
sets the tooltip in the constructor using `getTooltip()`, which returns `null` for instance-based classes since their
fields have not been initialized yet. This is a bad pattern, but we got away with when keys were all hardcoded in the
methods themselves.