AbstractDrawingLikeTools are now more careful about mouse presses and drags
Mouse presses and drags have to agree with each other. But also mouse drags should be treated like mouse moves with the
additional possibility of drawing the freehand options. This is required because a canceled map drag is still a drag as
far as Swing is concerned, and it will not generate `mouseMoved()` while a drag is still occurring.