Clean up WallTopologyRig
Its method now only accept `MovableWall` and `MovableVertex`, not the underlying `Wall` or `Vertex`. This makes it more
consistent and a little easier for callers to use it.

This also resolves a potential issue with vertex merging. Merging the two vertices of an standalone wall results in the
wall and the two vertices being deleted. To represent this possibility, the result is now wrapped in an `Optional<>`,
with an empty result indicating there was no survivor.