Remove change to consistently use LinkedHashMap
The change to `LinkedHashMap<>` should not cause any issues, even with serialization, since we copy the map contents
into our chosen implementation anyways. However, it does mean that new campaigns files with have `<linked-hash-map>` as
the values under `<lightSourcesMap>`, where they used to have `<map>`. So to avoid any possible issues, and to meet the
goal of zero campaign file changes in 1.18, we're going back to consistent `HashMap<>` for light source maps.

The ubiquitous copying of light source maps remains.