Relayout config panel
With both unique light sources and unique auras, the panel is a bit bigger and not everything was set up well on the
grid in order to stretch. The text fields for unique light sources and unique auras are once again inside a
`JScrollPane` with a larger preferred height so that edit is less painful.

Also removes empty rows and the pack of hspacers that was at the top of the config panel.