Add support for walls in UVTT importer
If the user requests walls (either via preference or prompt), then walls are generated to represent the UVTT line of
sight information. No VBL or MBL is generated in this case. If the user instead requests masks, then the legacy VBL+MBL
is generated.

If the user cancels the prompt for which type to use, the entire import is cancelled.