Add bounds to numeric preferences
The bounds are housed in a new `Preference` subclass (`NumericPreference`) that is specialized for numeric
types. The new method `Preference#constrain(T)` is used to clamp values to valid ranges before storing and after
loading. `NumericPreference` overrides this to do the clamping, while the base implementation just returns the original
value.

Using these bounds allows us to remove some redundant code:
- `ByteType` is no more than `IntegerType` constrained to the range [0, 255]. So now we just use `IntegerType` with
  those bounds specified.
- Preference validators aren't needed anymore since they were only used to enforce a positive frame rate cap. That is
  now handled more cleanly by added a lower bound on the frame rate cap.

Also removed is is `FloatType` class since the single `Float` preference is only used in a `Double` context, and the
value could just be casted down if it is ever needed in a `Float` context.