Wall control panel no longer syncs updates to clients
This responsibility is better placed in the Wall topology tool itself.

With this change, the control panel only has to manage the immutable `Wall.Data` rather than an entire `Wall`. This also
cleans the panel up a bit since it now longer has to keep track of the current zone either.