Move various bespoke actions into AppActions
Several actions existed both as part of the `TokenPopupMenu` as well as `PointerTool`/`StampTool`, sometimes with small
inconsitencies in behaviour. These are now captured exclusively in the following `AppActions`:
- `SET_FACING_ACTION`
- `EXPOSE_VISIBLE_AREA_ACTION`
- `EXPOSE_VISIBLE_AREA_ONLY_ACTION`
- `EXPOSE_LAST_PATH_ACTION`
- `EXPOSE_ALL_OWNED_ACTION``

These also accounted for the only uses of `AppActions.menuShortcut` outside of `AppActions` itself, so the field is now
private. The few external callers that require the modifier key can use `withMenuShortcut()` to add CTRL or CMD to a
`KeyStroke`.