Fix fonts and other atlas textures to use premultiplied alpha
The texture atlas is premultiplied as part of the atlas build. Fonts, on the other hand, do not have a convenient
configuration to accomplish this, so instead we have to rewrite the font texture pixel-by-pixel when it is loaded.

Aside from that one downside, we can now consistently composite render layers using premultiplied alpha, and don't have
to switch to semi-traditional alpha compositing just for text.