Avoid using the `workImage` in ZoneRenderer#renderTokens()
At this point, it was only being used to calculate the new size of a isometric token, to set its width and
height. Leaving aside the objectionable nature of this logic being in the renderer, it's not necessary to perform a
series of image transformations just to calculate the final size.

This this, there is no more need for `flipImageMap`. The similarly used `flipIsoImageMap` still has a use in
`showBlockedMoves()`, so it can't be removed yet.