Significantly simplify FacingArrowRenderer
All arrows are similar to one another, so we only need to create one and then transform it into position and size. This
alleviates the need for a bunch of casework trying to detect which arrow to use, only to get the same result.

For square tokens, instead of adjusting the final translation, just calculate the distance to the point. The existing
rotation can be leveraged to just move in that direction until the point is where it needs to be. Same gets done for
other tokens, but the calculation is much simpler there.

Since we only have on arrow now (and they are so cheap to build anyways), there is no need to cache them in the
quiver. The quiver was a memory leak anyway since there is no way to expunge values, and the size is extremely sensitive
to equality (it's a double after all).