Clean up ZoneRenderer.renderTokens()
The clipping conditions for figures and alway-visible tokens are already checked by `isTokenInNeedOfClipping()`, so it
isn't necessary to repeat those when trying to call `tokenRenderer.renderToken()`. The special clips used for these
types have been moved up as well, so that we don't need any additional logic when it comes time to render the token.

The existing checks to make sure a token is on screen and visible are redundant as `ZoneViewModel` already checks
this. So we just leverage it to grab a `TokenPosition`, guaranteeing that the token is okay to render without having to
duplicate the logic.

There was some leftover logic that calculated the token's transformed bounds, but in screen space. This is wasteful
since `ZoneViewModel` already calculates that (though in world space). On top of that, it is only used to set the clip
for those figure and always-visible tokens that need clipping, yet the area was calculated for all tokens. Both these
wastes are avoided now.

This comes with a mild performance benefit. Essentially all of the time in `ZoneRenderer.renderTokens()` is spent on
`TokenRenderer.renderToken()` now, with the surrounding logic contributing very little overhead.