Refine how ZoneViewModel keeps track of tokens.
Previously I made a performance change for `ZoneViewModel` to only track the position of on screen tokens. This was
based on a midreading of the old `ZoneRenderer` code and comments that suggested `tokenLocationCache` did the same
thing. But this was not true, so `ZoneViewModel` has been restored to tracking all tokens on the map so that full
functionality is restored, particularly for rendering paths for tokens that other clients are moving. It turns out the
core performance issue was just the way tokens were rendered, and the fact the `ZoneViewModel` tracks all these tokens
was a red herring.

There is still a need to check if certain tokens are on-screen, and this is what the old `tokenLocationMap` used to
do. So `ZoneViewModel` now also maintains a set of tokens that are on screen but which are not guaranteed to be visible
to the player.

I've also moved the creation of `TokenPosition` from `Token` into a static factory on `TokenPosition`. It will soon be
useful to be able to create arbitrary `TokenPosition` outside of the main update look.