Fix performance of `ZoneRenderer.getTokenBounds()`
Until recently, it was a fairly cheap operation that did a map lookup and little more. Then it got change to loop over
all tokens in its layer looking for itself. This has been change to more closely reflect the original, by directly
looking up the token's `TokenPosition` and using that for the answer.

All add `getTokenIdsOnScreen()`, which is like `getTokensOnScreen()` but returns IDs and doesn't sort the tokens. It's
really just `getTokenInBounds()`, but doesn't require the caller to know how to build the current viewport rectangle.