Preferences dialog: Change Startup help text to per-field entries
Also clear up the labels and tooltips. Each field now has a separate label, tooltip, and help text. The label is short
and descriptive to humans (does not include JVM flags, for example). The tooltips describe the nuts and bolts (so they
do include JVM flags in some cases). The help text contains only what is needed for a user to not make immediate errors,
but does not repeat the information in the tooltip.