Add new preference storage utilitiy
The new `PreferenceStore` provides much the same functionality as `AppPreferences.Preference` and its associated
types. But `PreferenceStore` can wrap an arbitrary `java.util.prefs.Preferences` so that it can be used not only by
`AppPreferences`, but `DeveloperOptions` and potentially others as well.

Once a `PreferenceStore` is created, preferences can be defined on it using any of its various `define*()` to assign a
type, label, tooltip and bounds for a key. This returns a `Preference<T>` that can be used to look up and write the
preference in a type-safe way. The `PreferenceStore` will also keep track of all preferences that have been defined so
they can be accessed dynamically.