Avoid calling getTokenBounds()
Callers now grab the world space token bounds. If screen space is needed, callers transform it. If the bounds must be
on-screen, callers explicitly check that.

For light source icons, the `ZoneViewModel` now tracks the positions of light sources. Ths `ZoneRenderer` and
`GdxRenderer` can use this to draw the icons where they need to be.

With no more callers of `ZoneRenderer.getTokenBounds()`, it has been removed.