Encapsulate campaign sight and lights in specialized collections
`Lights` is a `Set<LightSource>` that guarantees unique IDs. `Sights` is a `Set<SightType>` that guarantees unique
names. Both of these provide a central place to decide the implementation used to store sights and lights, and they are
also easier to write and use than `Map<GUID, LightSource>` and `Map<String, SightType>`.

`CategorizedLights` is similar, but is not a true collection type, though it is similar to a `Map<String, Lights>`. It
is more convenient than writing `Map<String, Map<GUID, LightSource>>`, and has a more constrained set of operations.

For backwards compatibility, `CampaignProperties` still has the `Map<String, Map<GUID, LightSource>> lightSourcesMap`
and `Map<String, SightType> sightTypeMap` fields. When loading a campaign, these are automatically converted to
`CategorizedLights` and `Sights`, respectively. When saving the campaign, they get converted back to the raw maps. This
ensures that campaigns in 1.18 can still be loaded in 1.17.