Nightly builds: fix latest stable tag detection
Use Github CLI's built-in functionality to filter out drags and prereleases (not just nightlies), and be explicit about
the ordering requirements.

Note: we could also just filter for the `isLatest` tag, but that would not detect that real latest stable tag, only the
one that has been manually marked as latest.