Make lib: URIs for macros be handled consistently
The new `Library#getMTScriptMacroInfoForUriPath(String)` explicitly resolves macro paths (which have a leading `macro/`
for lib:tokens) so that `Library#getMTScriptMacroInfo(String)` is free to resolve pure macro names. This liberates
`LibraryToken#getMTScriptMacroInfo` from needing to parse full URIs.

Corresponding changes have been made to the line parser so that library namespace are handled consistently between URI
and library `MacroLocation`s, while permitting the appropriate macro resolution strategies to be used depending on which
kind of `MacroLocation` we are resolving.